+++ vllm_omni/diffusion/models/flux2/__init__.py
@@ -7,7 +7,6 @@ from vllm_omni.diffusion.models.flux2.flux2_transformer import (
 )
 from vllm_omni.diffusion.models.flux2.pipeline_flux2 import (
     Flux2Pipeline,
-    get_flux2_post_process_func,
 )
 from vllm_omni.diffusion.models.flux2.pipeline_flux2_klein import (
     Flux2KleinPipeline,
@@ -17,5 +16,4 @@ __all__ = [
     "Flux2Pipeline",
     "Flux2KleinPipeline",
     "Flux2Transformer2DModel",
-    "get_flux2_post_process_func",
 ]
